Rename dashboard post to "/dashboard/posts"
