+++ frontend/src/pages/dashboard/posts/_slug/Index.vue
